+++ envs/ant.py
@@ -44,7 +44,7 @@ class Ant(MujocoBase):
                 action,
                 self.action_bounds[ActionBoundsIdx.lb],
                 self.action_bounds[ActionBoundsIdx.ub],
-                dtype=np.float128
+                dtype=np.float64
             )
         else:
             clipped_action = action
