+++ envs/half_cheetah.py
@@ -53,7 +53,7 @@ class HalfCheetah(MujocoBase):
                 action,
                 self.action_bounds[ActionBoundsIdx.lb],
                 self.action_bounds[ActionBoundsIdx.ub],
-                dtype=np.float128
+                dtype=np.float64
             )
         else:
             clipped_action = action
