+++ policies/nn_policy.py
@@ -149,14 +149,14 @@ class DeepGaussian(NeuralNetworkPolicy):
             self.std_dev - self.std_decay,
             self.std_min,
             np.inf,
-            dtype=np.float128
+            dtype=np.float64
         )
 
     def draw_action(self, state) -> np.array:
         means = np.array(super().draw_action(state=state), dtype=np.float64)
-        # action = np.array(np.random.normal(means, self.std_dev), dtype=np.float128)
+        # action = np.array(np.random.normal(means, self.std_dev), dtype=np.float64)
         action = np.array(
             means + self.std_dev * np.random.normal(0, 1, self.dim_action),
-            dtype=np.float128
+            dtype=np.float64
         )
         return action
